# HubBancario/API/OpenApi/hub-bancario-api.yaml
openapi: 3.0.3
info:
  title: Hub Bancário API
  version: "1.0.0"
  description: |
    API B2B para Integração Pix e Faturamento. 
    Este contrato atende aos fluxos completos de Contas, Chaves, Faturamento e emissão de cobranças do Banco Central via Open Finance.
  contact:
    name: Squad 09 - Hub Bancário
servers:
  - url: https://localhost:5001
    description: Ambiente de Desenvolvimento (HTTPS)
  - url: http://localhost:5000
    description: Ambiente de Desenvolvimento (HTTP)

# Configuração Global de Segurança (JWT)
security:
  - BearerAuth: []

paths:
  /api/accounts:
    post:
      summary: Cadastra uma nova conta lojista
      tags: [Accounts]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAccountRequest'
      responses:
        "201":
          description: Conta criada com sucesso
        "400":
          description: Falha na validação dos dados
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'

  /api/accounts/{id}:
    put:
      summary: Atualiza os dados bancários de uma conta
      tags: [Accounts]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAccountRequest'
      responses:
        "204":
          description: Conta atualizada com sucesso
        "400":
          description: Inconsistência nos IDs enviados
    delete:
      summary: Realiza a exclusão lógica da conta
      tags: [Accounts]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Conta inativada com sucesso

  /api/pix-keys:
    post:
      summary: Vincula uma nova chave Pix a uma conta
      tags: [Pix Keys]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePixKeyRequest'
      responses:
        "201":
          description: Chave vinculada com sucesso
        "409":
          description: Conflito - Chave já cadastrada

  /api/pix-keys/{id}:
    delete:
      summary: Desvincula uma chave Pix
      tags: [Pix Keys]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Chave removida com sucesso

  /api/client-secrets/{id}/revoke:
    post:
      summary: Revoga imediatamente uma credencial de segurança
      tags: [Client Secrets]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Credencial inativada para acesso

  /api/invoices:
    post:
      summary: Emite uma nova fatura (intenção de cobrança)
      tags: [Invoices]
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
          description: Chave para evitar duplicidade de faturamento
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInvoiceRequest'
      responses:
        "201":
          description: Fatura criada
          
  /api/pix-charges:
    post:
      summary: Emite uma cobrança Pix (Gera o QR Code EMV)
      tags: [Pix Charges]
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePixChargeRequest'
      responses:
        "201":
          description: Cobrança gerada com sucesso no banco parceiro
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PixChargeResponse'

  /api/pix-charges/{txId}/status:
    get:
      summary: Consulta o status de uma cobrança gerada
      tags: [Pix Charges]
      parameters:
        - name: txId
          in: path
          required: true
          schema:
            type: string
            maxLength: 35
      responses:
        "200":
          description: Status retornado com sucesso (Active, Paid, Expired)

  /api/webhooks/pix:
    post:
      summary: Recebe as notificações de pagamento do Banco
      tags: [Webhooks (Callbacks)]
      security: [] # Esta rota tem política de segurança própria (IP whitelist/mTLS)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object # O formato varia conforme o banco, deixamos em aberto
      responses:
        "202":
          description: Evento recebido e enfileirado para processamento assíncrono

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    ProblemDetails:
      type: object
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string

    CreateAccountRequest:
      type: object
      required: [document, bankId, accountNumber, agency]
      properties:
        secretId:
          type: string
          format: uuid
        document:
          type: string
          example: "12345678909"
        bankId:
          type: string
          example: "341"
        accountNumber:
          type: string
          example: "12345-6"
        agency:
          type: string
          example: "0001"

    UpdateAccountRequest:
      type: object
      properties:
        id:
          type: string
          format: uuid
        bankId:
          type: string
        accountNumber:
          type: string
        agency:
          type: string

    CreatePixKeyRequest:
      type: object
      required: [keyValue, accountId]
      properties:
        keyValue:
          type: string
          example: "loja@meuerp.com.br"
        accountId:
          type: string
          format: uuid

    CreateInvoiceRequest:
      type: object
      required: [accountId, amount, dueDate]
      properties:
        accountId:
          type: string
          format: uuid
        amount:
          type: number
          format: double
          example: 150.75
        dueDate:
          type: string
          format: date
          example: "2026-05-30"
        externalReference:
          type: string
          example: "PEDIDO-9988"

    CreatePixChargeRequest:
      type: object
      required: [invoiceId, chargeType]
      properties:
        invoiceId:
          type: string
          format: uuid
        chargeType:
          type: string
          example: "COB"
          description: "Tipos suportados: COB (Imediata) ou COBV (Com Vencimento)"

    PixChargeResponse:
      type: object
      properties:
        txId:
          type: string
        emv:
          type: string
          description: O código Pix Copia e Cola
        qrCodeBase64:
          type: string
        status:
          type: string